Move client functionality to the `ClientApp` in the ``client` module.
